all endpoints manually tested
