+++ backend/ASTU-backend-group-5/internal/http/handlers/account/account_handlers.go
@@ -1,4 +1,4 @@
-package handlers
+package account
 
 import (
 	"blogApp/internal/domain"
