+++ backend/ASTU-backend-group-5/internal/http/handlers/token/token_handlers.go
@@ -1,4 +1,4 @@
-package handlers
+package token
 
 import (
 	"blogApp/internal/domain"
@@ -7,6 +7,7 @@ import (
 	"net/http"
 	"strings"
 	"time"
+
 	"github.com/gin-gonic/gin"
 )
 
@@ -53,11 +54,10 @@ func (h *TokenHandler) LogOut(c *gin.Context) {
 	token := domain.Token{}
 	token.AccessToken = strings.Split(c.GetHeader("x_access_token"), " ")[1]
 	token.RefreshToken = strings.Split(c.GetHeader("x_refresh_token"), " ")[1]
-	if token.AccessToken == "" || token.RefreshToken  == "" {
+	if token.AccessToken == "" || token.RefreshToken == "" {
 		c.JSON(http.StatusBadRequest, gin.H{"error": "Authorization token required"})
 		return
 	}
-
 
 	if err := c.ShouldBindJSON(&token); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
